pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "litellm-memory"
version = "0.1.0"
description = "memory proxy"
authors = [
    {name = "zeddq",email = "zeddq1@gmail.com"}
]
readme = "README.md"
requires-python = "=3.13.8"
dependencies = [
    "fastapi>=0.115.5",
    "httpx~=0.28.1",
    "litellm-types (>=0.0.16,<0.0.17)",
    "litellm[proxy] (>=1.79.0,<2.0.0)",
    "opentelemetry-api (>=1.38.0,<2.0.0)",
    "opentelemetry-distro (>=0.59b0,<0.60)",
    "opentelemetry-exporter-otlp (>=1.38.0,<2.0.0)",
    "opentelemetry-sdk (>=1.38.0,<2.0.0)",
    "opentelemetry-instrumentation-logging (>=0.45b0)",
    "opentelemetry-instrumentation-fastapi (>=0.45b0)",
    "prisma (>=0.15.0,<0.16.0)",
    "pyyaml~=6.0.3",
    "supermemory (>=3.4.0,<4.0.0)",
    "uvicorn>=0.29.0"
]

# Note: LiteLLM is required as a CLI tool, not as a Python package dependency.
# Install separately: pip install litellm
# This project uses the 'litellm' binary via subprocess instead of SDK imports.

[dependency-groups]
dev = [
    "deptry>=0.23.1",
    "pip-check>=3.2.1",
    "debugpy (>=1.8.17,<2.0.0)",
    "notebook (>=7.4.7,<8.0.0)",
    "openai (>=2.6.0,<3.0.0)",
    "pydevd-pycharm (>=252.26830.99,<252.26831.0)",
    "debugpy (>=1.8.17,<2.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "mypy (>=1.18.2,<2.0.0)",
]
test = [
    "pytest (>=8.4.2)",
    "pytest-asyncio (>=1.2.0)",
    "pytest-cov (>=7.0.0)",
    "pytest-mock (>=3.15.1)",
    "pytest-timeout (>=2.4.0)",
    "pytest-check (>=2.6.0,<3.0.0)",
    "psutil (>=7.1.3,<8.0.0)",
    "pytest-xdist (>=3.8.0,<4.0.0)",
]

[tool.poetry]
packages = [
    {include = "proxy", from = "src"},
    {include = "interceptor", from = "src"},
    {include = "telemetry", from = "src"}
]

[tool.poetry.scripts]
start-proxies = "start_proxies:main"
interceptor = "interceptor.cli:main"

[tool.pytest.ini_options]
# Pytest configuration
minversion = "8.0"
testpaths = [".", "tests/src"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "--tb=short",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

# Markers for categorizing tests
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
    "real_api: marks tests that require real API keys and make actual API calls",
]

# Timeout for tests (in seconds)
#timeout = 30

[tool.coverage.run]
# Coverage configuration
source = ["."]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/virtualenv/*",
    ".venv/*",
]
branch = true

[tool.coverage.report]
# Coverage report configuration
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"
